NetworkConfigurator: calculate mac address closer to the usage
