+++ pkg/network/infraconfigurators/macvtap.go
@@ -6,6 +6,7 @@ import (
 
 	"github.com/vishvananda/netlink"
 
+	v1 "kubevirt.io/client-go/api/v1"
 	"kubevirt.io/client-go/log"
 	"kubevirt.io/kubevirt/pkg/network/cache"
 	netdriver "kubevirt.io/kubevirt/pkg/network/driver"
@@ -15,15 +16,16 @@ import (
 type MacvtapPodNetworkConfigurator struct {
 	podInterfaceName string
 	podNicLink       netlink.Link
+	vmiSpecIface     *v1.Interface
 	vmMac            *net.HardwareAddr
 	launcherPID      int
 	handler          netdriver.NetworkHandler
 }
 
-func NewMacvtapPodNetworkConfigurator(podIfaceName string, vmMac *net.HardwareAddr, handler netdriver.NetworkHandler) *MacvtapPodNetworkConfigurator {
+func NewMacvtapPodNetworkConfigurator(podIfaceName string, vmiSpecIface *v1.Interface, handler netdriver.NetworkHandler) *MacvtapPodNetworkConfigurator {
 	return &MacvtapPodNetworkConfigurator{
 		podInterfaceName: podIfaceName,
-		vmMac:            vmMac,
+		vmiSpecIface:     vmiSpecIface,
 		handler:          handler,
 	}
 }
@@ -45,7 +47,7 @@ func (b *MacvtapPodNetworkConfigurator) preparePodNetworkInterface() error {
 
 func (b *MacvtapPodNetworkConfigurator) generateDomainIfaceSpec() api.Interface {
 	return api.Interface{
-		MAC: &api.MAC{MAC: b.podIfaceMAC()},
+		MAC: &api.MAC{MAC: b.vmMac},
 		MTU: &api.MTU{Size: strconv.Itoa(b.podNicLink.Attrs().MTU)},
 		Target: &api.InterfaceTarget{
 			Device:  b.podNicLink.Attrs().Name,
@@ -61,15 +63,15 @@ func (b *MacvtapPodNetworkConfigurator) DiscoverPodNetworkInterface(podIfaceName
 	}
 	b.podNicLink = link
 
-	return nil
-}
-
-func (b *MacvtapPodNetworkConfigurator) podIfaceMAC() string {
-	if b.vmMac != nil {
-		return b.vmMac.String()
-	} else {
-		return b.podNicLink.Attrs().HardwareAddr.String()
+	b.vmMac, err = retrieveMacAddressFromVMISpecIface(b.vmiSpecIface)
+	if err != nil {
+		return err
 	}
+	if b.vmMac == nil {
+		b.vmMac = &b.podNicLink.Attrs().HardwareAddr
+	}
+
+	return nil
 }
 
 func (b *MacvtapPodNetworkConfigurator) PreparePodNetworkInterface() error {
@@ -78,7 +80,7 @@ func (b *MacvtapPodNetworkConfigurator) PreparePodNetworkInterface() error {
 
 func (b *MacvtapPodNetworkConfigurator) GenerateDomainIfaceSpec() api.Interface {
 	return api.Interface{
-		MAC: &api.MAC{MAC: b.podIfaceMAC()},
+		MAC: &api.MAC{MAC: b.vmMac},
 		MTU: &api.MTU{Size: strconv.Itoa(b.podNicLink.Attrs().MTU)},
 		Target: &api.InterfaceTarget{
 			Device:  b.podNicLink.Attrs().Name,
